support moving cursor to beginning and end of line.

+++ include/astedit/textedit.h
@@ -26,7 +26,7 @@ int read_from_textedit(struct TextEdit *edit, int offset, char *dstBuffer, int s
 
 
 void erase_from_textedit(struct TextEdit *edit, int offset, int length);
-void insert_codepoint_into_textedit(struct TextEdit *edit, unsigned long codepoint);
+void insert_codepoint_into_textedit(struct TextEdit *edit, uint32_t codepoint);
 
 /* fill TextEdit with some text loaded from a file. For debugging purposes. */
 void textedit_test_init(struct TextEdit *edit, const char *filepath);
